Completed Task 1
I created a singelton object called ProductsWarehouse. In this object will be saved productslist
