+++ Advanced Kotlin/LittleLemonMenu/app/src/main/java/com/littlelemon/menu/MainActivity.kt
@@ -9,27 +9,11 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.core.view.MenuCompat
+import com.littlelemon.menu.ProductsWarehouse.productsList
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.update
 
 class MainActivity : ComponentActivity() {
-
-    val productsList = mutableListOf(
-        ProductItem("Black tea", 3.00, "Drinks", R.drawable.black_tea),
-        ProductItem("Green tea", 3.00, "Drinks", R.drawable.green_tea),
-        ProductItem("Espresso", 5.00, "Drinks", R.drawable.espresso),
-        ProductItem("Cappuccino", 8.00, "Drinks", R.drawable.cappuccino),
-        ProductItem("Latte", 8.00, "Drinks", R.drawable.latte),
-        ProductItem("Mocha", 10.00, "Drinks", R.drawable.mocha),
-        ProductItem("Boeuf bourguignon", 15.00, "Food", R.drawable.boeuf_bourguignon),
-        ProductItem("Bouillabaisse", 20.00, "Food", R.drawable.bouillabaisse),
-        ProductItem("Lasagna", 15.00, "Food", R.drawable.lasagna),
-        ProductItem("Onion soup", 12.00, "Food", R.drawable.onion_soup),
-        ProductItem("Salmon en papillote", 25.00, "Food", R.drawable.salmon_en_papillote),
-        ProductItem("Quiche Lorraine", 17.00, "Dessert", R.drawable.quiche_lorraine),
-        ProductItem("Custard tart", 14.00, "Dessert", R.drawable.custard_tart),
-        ProductItem("Croissant", 7.00, "Dessert", R.drawable.croissant),
-    )
 
     private val productsState: MutableStateFlow<Products> =
         MutableStateFlow(Products(productsList))
